+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/ChannelMediaGroupMessage.kt
@@ -1,0 +1,22 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.message
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.MediaGroupIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
+import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaGroupMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MediaGroupContent
+import com.soywiz.klock.DateTime
+
+data class ChannelMediaGroupMessage(
+    override val messageId: MessageIdentifier,
+    override val chat: Chat,
+    override val date: DateTime,
+    override val mediaGroupId: MediaGroupIdentifier,
+    override val content: MediaGroupContent,
+    override val editDate: DateTime?,
+    override val forwarded: ForwardedMessage?,
+    override val replyTo: Message?,
+    override val replyMarkup: InlineKeyboardMarkup?
+) : MediaGroupMessage