+++ src/jvmMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/UserLocale.kt
@@ -1,0 +1,7 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types
+
+import java.util.*
+
+fun User.javaLocale(): Locale? = languageCode ?.let {
+    Locale.forLanguageTag(it)
+}