+++ src/jvmMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/StringFormatActual.kt
@@ -1,0 +1,5 @@
+package com.github.insanusmokrassar.TelegramBotAPI.utils
+
+import kotlin.text.format
+
+actual fun String.format(vararg args: Any?): String = format(*args)